feat(mobile): add auth
