+++ apps/mobile/src/routes/tab-routes.tsx
@@ -1,9 +1,15 @@
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
-import { getFocusedRouteNameFromRoute } from '@react-navigation/native';
-import { useCustomModal } from '../context';
+import {
+  getFocusedRouteNameFromRoute,
+  useNavigation,
+} from '@react-navigation/native';
+import { useCustomModal, useUserContext } from '../context';
 import TodoScreen from '../screens/todo-screen';
 import TodoListStack from './todo-routes';
 import { MaterialIcons } from '@expo/vector-icons';
+import { getAuth } from 'firebase/auth';
+import { api } from '@todo/commons';
+import AsyncStorage from '@react-native-async-storage/async-storage';
 const Tab = createBottomTabNavigator();
 
 const TabBarIcon = ({ route, color }: { route: string; color: string }) => {
@@ -15,12 +21,14 @@ const TabBarIcon = ({ route, color }: { route: string; color: string }) => {
         <MaterialIcons name="add-circle-outline" size={24} color={color} />
       );
     default:
-      return <>Unknow Route</>;
+      return <MaterialIcons name="logout" />;
   }
 };
 
 export function Tabs() {
   const [, dispatch] = useCustomModal();
+  const navigation = useNavigation();
+  const [user, removeUser] = useUserContext();
 
   return (
     <Tab.Navigator
@@ -65,6 +73,28 @@ export function Tabs() {
           tabPress: (e) => {
             e.preventDefault();
             dispatch(true);
+          },
+        }}
+      />
+      <Tab.Screen
+        name="Todo/Settings"
+        component={() => null}
+        options={({ route }) => ({
+          tabBarIcon: ({ focused, color }) => (
+            <TabBarIcon route={route.name} color={color} />
+          ),
+        })}
+        listeners={{
+          tabPress: async (e) => {
+            e.preventDefault();
+            try {
+              await getAuth().signOut();
+              await api.users.revokeCustomToken(user?.id as string);
+              await AsyncStorage.removeItem('sessionToken');
+              removeUser(null)
+            } catch (err) {
+              console.log(err);
+            }
           },
         }}
       />
