+++ apps/mobile/src/screens/auth/signup-screen.tsx
@@ -6,6 +6,10 @@ import { Dimensions, Image, Pressable } from 'react-native';
 import styled from 'styled-components/native';
 import { CustomSafeAreaView } from '../../components/safe-area-view';
 import { SignupFormView } from '../../components/signup-form-view';
+import {
+  createUserWithEmailAndPassword,
+} from 'firebase/auth';
+import { auth } from '../../utils/firebase';
 
 const WIDTH = Dimensions.get('window').width;
 const HEIGHT = Dimensions.get('window').height;
@@ -36,11 +40,23 @@ export const SignupScreen = () => {
   async function handleSignup<T>(data: Partial<User & { [key: string]: any }>) {
     const { email, username, password } = data;
 
-    await api.users.createUser({
-      email: email as string,
-      username: username as string,
-      password: password as string,
-    });
+    try {
+      const result = await createUserWithEmailAndPassword(
+        auth,
+        email as string,
+        password as string
+      );
+
+      const user = await api.users.createUser({
+        username: username as string,
+        email: email as string,
+        id: result.user.uid
+      })
+
+      console.log(user);
+    } catch (err) {
+      console.log(err);
+    }
   }
 
   return (
